Add files via upload
Analysis of 3 supermarket branches during 3 months, between January and March 2019. The data was uploaded from the Kaggle (https://www.kaggle.com/datasets/aungpyaeap/supermarket-sales)):

overall, 5510 products were sold and for that, the supermarket received a total of $322.970.
On average, each client spent $323 and the average rating was 6.97/10.
The gender of the customers is well divided, 50.1% woman and 49.9% man.
35% of the income was paid with cash, 34% with Ewallet and 31% with credit card.
'Food and beverages' is the product line that most returns money, followed by 'sports and travel', 'Electronic accessories', 'Fashion accessories', 'Home and lifestyle' and 'Health and beauty'. The difference between the first to the last is 2.1%, or $6951.
'Eletronic accessories' are the most sold ones (971 products). 'Health and beauty is the least sold (854 products).
March 09th was the day with the biggest revenue, $7407,05 and February 13th the least one, with $934,24 in all branches.
Store A:

sold 1859 products and earned $106.200,00. Each client spent on average $312 and the rating was 7.03/10.
Ewallet was used to pay more money than the other payments (37% and 31% each).
'Home and lifestyle' was the product line most profitable. Food and beverage at this branch stayed in 4th position.
Store B:

sold 1820 products, earning also $106.200,00. The clients spend on average $320.00, with rating of 6.82/10.
At this store, more money was billed with credit card.
'Sports and travel' was the most profitable and 'Food and beverage' the least.
Store C:

sold 1831 products, earned $110.570,00. Customer rating was around 7.07/10 and each spend on average $337.
at this store there are more woman buying than men (54% to 46%, respectivelly)
cash was the way more money came in than the other options.
'Food and beverage' was the product line most profitable, followed by 'Fashion accessories'.
Analysing just what men spent on Store B, for the period, they spent on average $313,35. On the 28th of February, only $63,32 and on 31 th of January $2601,68. They represent at store B 51.2% of the public, the billing payment type is well divided and the product line where they most spent money is 'Health and beauty'.